+++ test/tst_dict_mem2.f90
@@ -32,7 +32,7 @@ contains
   subroutine mem_rem(dealloc)
     logical, intent(in) :: dealloc
     type(dictionary_t) :: d
-    real(dp) :: va(400,400) ! roughly 1.22 MB
+    real(real64) :: va(400,400) ! roughly 1.22 MB
     type(variable_t) :: v
     va = 0.
     call extend(d,'hello'.kv.va)
